% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aic_matrices_shrunk_observation.R
\name{aic.given.radius.and.lag.shrunk}
\alias{aic.given.radius.and.lag.shrunk}
\title{Function for making a new AIC matrix given DNS data for a given radius and lag}
\usage{
aic.given.radius.and.lag.shrunk(
  r_and_h_untransformed,
  vortices.pp,
  scars.pp,
  sDiv,
  indices,
  radiiForPrint,
  startLag,
  radiiModulo = 1000L,
  xlim,
  ylim
)
}
\arguments{
\item{r_and_h_untransformed}{r_and_h_combined in single number}

\item{vortices.pp}{votrices as spatstat point process}

\item{scars.pp}{scars as spatstat point process}

\item{sDiv}{horisontal divergence}

\item{indices}{which indices to use}

\item{radiiForPrint}{which radii to print}

\item{startLag}{First lag, used to decide when to print the radius}

\item{radiiModulo}{Parameter for how to decode the r_and_h_untransformed to r and h}

\item{xlim}{how much to shrink the window in the x direction}

\item{ylim}{how much to shrink the window in the y direction}
}
\description{
Function for making a new AIC matrix given DNS data for a given radius and lag
}
