% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingFunctions.R
\name{aicPlot}
\alias{aicPlot}
\title{Function for plotting a snapshot of surface field and surface pp together}
\usage{
aicPlot(aicMatrix, lagValues, timeScale, lengthScale, radiar)
}
\arguments{
\item{aicMatrix}{Matrix with surface field data for the interesting snapshot}

\item{lagValues}{Coordinates for the point process}

\item{timeScale}{how many timescales per frame?}

\item{lengthScale}{how many lengthscales per pixle?}

\item{radiar}{description}
}
\description{
Function for plotting a snapshot of surface field and surface pp together
}
