% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calculateLocVar_periodic}
\alias{calculateLocVar_periodic}
\title{Calculates the local variance with periodic boundary conditions}
\usage{
calculateLocVar_periodic(inWin, inShell, insDiv, inDims)
}
\arguments{
\item{inWin}{Window}

\item{inShell}{Shell of window}

\item{insDiv}{field of which the local variance should be calculated}

\item{inDims}{Integer vector of dimensions for input, should be <xDim, yDim, winDim, shellDim>}
}
\value{
Vector with the local variance
}
\description{
Calculates the local variance with periodic boundary conditions
}
