% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point_tracking.R
\name{closeness.of.points}
\alias{closeness.of.points}
\title{Function for finding the distance between all points at one timestep, and all points at the next timestep}
\usage{
closeness.of.points(
  points.t1,
  points.t2,
  weight_dist = 256,
  moveOnTorus = TRUE
)
}
\arguments{
\item{points.t1}{dataframe with points at the first of the two timesteps}

\item{points.t2}{dataframe with points at the second of the two timesteps}

\item{weight_dist}{parameter to indicate how to weight difference in distance between to points vs the difference in size of two points}

\item{moveOnTorus}{set to TRUE if points move on a torus}
}
\description{
Function for finding the distance between all points at one timestep, and all points at the next timestep
}
