% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aic_matrices.R
\name{combine.covariates.betasq.dns}
\alias{combine.covariates.betasq.dns}
\title{Function for creating beta squared covariates for the desired indices for DNS data}
\usage{
combine.covariates.betasq.dns(sDiv, indices, h)
}
\arguments{
\item{sDiv}{horrisontal divergence complete array}

\item{indices}{indices to use points from}

\item{h}{lag}
}
\description{
Function for creating beta squared covariates for the desired indices for DNS data
}
