% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aic_matrices_shrunk_observation.R
\name{combine.covariates.betasq.dns.shrunk}
\alias{combine.covariates.betasq.dns.shrunk}
\title{Function for creating beta squared covariates for the desired indices for DNS data}
\usage{
combine.covariates.betasq.dns.shrunk(sDiv, indices, h, xlim, ylim)
}
\arguments{
\item{sDiv}{horrisontal divergence complete array}

\item{indices}{indices to use points from}

\item{h}{lag}

\item{xlim}{how much to shrink the window in the x direction}

\item{ylim}{how much to shrink the window in the y direction}
}
\description{
Function for creating beta squared covariates for the desired indices for DNS data
}
