% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aic_matrices.R
\name{combine.covariates.betasq.experimental}
\alias{combine.covariates.betasq.experimental}
\title{Function for creating the beta squared covariates for experimental data}
\usage{
combine.covariates.betasq.experimental(hDiv, indices, ensembles, h, skim = 0)
}
\arguments{
\item{hDiv}{horrisontal divergence complete array}

\item{indices}{indices to use points from}

\item{ensembles}{vector with all ensembles to use}

\item{h}{lag}

\item{skim}{how many pixles to skim off the edge?}
}
\description{
Function for creating the beta squared covariates for experimental data
}
