% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aic_matrices.R
\name{combine.covariates.dns}
\alias{combine.covariates.dns}
\title{Function for creating local variance covariates for the desired indices for DNS data}
\usage{
combine.covariates.dns(sDiv, indices, r, h)
}
\arguments{
\item{sDiv}{horrisontal divergence complete array}

\item{indices}{indices to use points from}

\item{r}{radius to use in calculating local variance}

\item{h}{lag}
}
\description{
Function for creating local variance covariates for the desired indices for DNS data
}
