% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aic_matrices.R
\name{combine.covariates.experimental}
\alias{combine.covariates.experimental}
\title{Function for creating covariates for the desired indices}
\usage{
combine.covariates.experimental(hDiv, indices, ensembles, r, h, skim = 0)
}
\arguments{
\item{hDiv}{horrisontal divergence complete array}

\item{indices}{indices to use points from}

\item{ensembles}{vector with all ensembles to use}

\item{r}{radius to use in calculating local variance}

\item{h}{lag}

\item{skim}{how many pixles to skim off the edge?}
}
\description{
Function for creating covariates for the desired indices
}
