% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aic_matrices.R
\name{combine.covariates.matrices}
\alias{combine.covariates.matrices}
\title{Function for combining covariates from multiple snapshots into one single image}
\usage{
combine.covariates.matrices(covariates, xdim, ydim)
}
\arguments{
\item{covariates}{list of all covariates to combine. Every element should be a matrix of the same dimension}

\item{xdim}{xdim of covariates}

\item{ydim}{ydim of covariates}
}
\description{
Function for combining covariates from multiple snapshots into one single image
}
