% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point_tracking.R
\name{dist.on.torus}
\alias{dist.on.torus}
\title{Function for distance between two points between 0 and 1 on the torus}
\usage{
dist.on.torus(x1, x2)
}
\arguments{
\item{x1}{First number}

\item{x2}{Second number}
}
\description{
Function for distance between two points between 0 and 1 on the torus
}
