% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distanceOnTorus.R
\name{distanceOnTorus}
\alias{distanceOnTorus}
\title{Distance between two points on a flat torus of arbitrary dimension}
\usage{
distanceOnTorus(d, x, y)
}
\arguments{
\item{d}{The dimension of the torus}

\item{x}{The first point}

\item{y}{The second point}
}
\value{
The distance between x and y.
}
\description{
Distance between two points on a flat torus of arbitrary dimension
}
\examples{
d <- 1
x <- 0.1
y <- 0.2
distanceOnTorus(d, x, y)

}
