% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surfaceDivergenceManipulation.R
\name{estimate.gradient.norm.sdiv}
\alias{estimate.gradient.norm.sdiv}
\title{Functions used for manipulation of the surface divergence
Function for estimating the norm of the gradient field of the surface divergence}
\usage{
estimate.gradient.norm.sdiv(sDiv, xdim, ydim, hx, hy)
}
\arguments{
\item{sDiv}{surface divergence}

\item{xdim}{Dimension in the x direction}

\item{ydim}{Dimension in the y direction}

\item{hx}{Step size in x-direction}

\item{hy}{Step size in y-direction}
}
\value{
Matrix with the norm of the gradient of the surface divergence
}
\description{
Functions used for manipulation of the surface divergence
Function for estimating the norm of the gradient field of the surface divergence
}
