% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grfPeriodicFunctions.R
\name{estimate.param.periodic.grf.pcprior}
\alias{estimate.param.periodic.grf.pcprior}
\title{Function to estimate parameters for a periodic GRF, using PC prior}
\usage{
estimate.param.periodic.grf.pcprior(
  observedData,
  priorSigma = c(1, 1),
  priorRange = c(1, 1),
  spde = NULL,
  torus = NULL,
  A = NULL,
  femMatrices = NULL,
  reducedOutput = FALSE
)
}
\arguments{
\item{observedData}{data to use for estimation}

\item{priorSigma}{Prior parameters for sigma}

\item{priorRange}{Prior parameters for the range}

\item{spde}{inla.spde2 object}

\item{torus}{Mesh}

\item{A}{A-matrix}

\item{femMatrices}{FEM approximation matrices}

\item{reducedOutput}{Set to TRUE if you want only summary.hyperpar and marginal.hyperpar}
}
\value{
INLA output from estimation of grf parameters
}
\description{
Function to estimate parameters for a periodic GRF, using PC prior
}
