% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grfPeriodicFunctions.R
\name{estimate.param.periodic.grf}
\alias{estimate.param.periodic.grf}
\title{Function to estimate parameters for a periodic GRF}
\usage{
estimate.param.periodic.grf(
  observedData,
  priorTheta = c(1, 1, 1),
  spde = NULL,
  torus = NULL,
  A = NULL,
  femMatrices = NULL,
  reducedOutput = FALSE
)
}
\arguments{
\item{observedData}{data to use for estimation}

\item{priorTheta}{Parameters for prior}

\item{spde}{inla.spde2 object}

\item{torus}{Mesh}

\item{A}{A-matrix}

\item{femMatrices}{FEM approximation matrices}

\item{reducedOutput}{Set to TRUE if you want only summary.hyperpar and marginal.hyperpar}
}
\value{
INLA output from estimation of grf parameters
}
\description{
Function to estimate parameters for a periodic GRF
}
