% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aic_matrices.R
\name{expand.aic.matrix.dns}
\alias{expand.aic.matrix.dns}
\title{Function for expanding an AIC matrix given dns data}
\usage{
expand.aic.matrix.dns(
  gamleRadiar,
  nyeRadiar,
  lags,
  gamleMatriser,
  vortices.pp,
  scars.pp,
  sDiv,
  indices,
  printProgress = TRUE,
  transformation = "identity"
)
}
\arguments{
\item{gamleRadiar}{Vector with all radii for local variance in the ole AIC matrices}

\item{nyeRadiar}{Vector with radii for local variance in expansion}

\item{lags}{Vector with all lag values to use}

\item{gamleMatriser}{list with the AIC matrices that should be expanded}

\item{vortices.pp}{votrices as spatstat point process}

\item{scars.pp}{scars as spatstat point process}

\item{sDiv}{horisontal divergence}

\item{indices}{which indices to use}

\item{printProgress}{Boolean variable to indicate whether to print progress or not}

\item{transformation}{Should the covariate be transformed, if value is "log" then a log transformation is applied}
}
\description{
Function for expanding an AIC matrix given dns data
}
