% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aic_matrices.R
\name{expand.aic.matrix.experimental}
\alias{expand.aic.matrix.experimental}
\title{Function for expanding an AIC matrix given experimental data}
\usage{
expand.aic.matrix.experimental(
  gamleRadiar,
  nyeRadiar,
  hmax,
  gamleMatriser,
  vortices.pp,
  scars.pp,
  hDiv,
  indices,
  ensembles,
  printProgress = FALSE,
  skim = 0
)
}
\arguments{
\item{gamleRadiar}{Vector with all radii for local variance in the ole AIC matrices}

\item{nyeRadiar}{Vector with radii for local variance in expansion}

\item{hmax}{Maximal lag}

\item{gamleMatriser}{list with the AIC matrices that should be expanded}

\item{vortices.pp}{votrices as spatstat point process}

\item{scars.pp}{scars as spatstat point process}

\item{hDiv}{horisontal divergence}

\item{indices}{which indices to use}

\item{ensembles}{vector of ensembles to use}

\item{printProgress}{Boolean variable to indicate whether to print progress or not}

\item{skim}{number of pixles to skim of the edge}
}
\description{
Function for expanding an AIC matrix given experimental data
}
