% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findVorticesCentroids.R
\name{explore.scar}
\alias{explore.scar}
\title{Function used to find scar properties given all the pixels in the scar}
\usage{
explore.scar(pixels, dims, np, skimage)
}
\arguments{
\item{pixels}{Data Frame with coordinates of all the pixels in the scar}

\item{dims}{vector of dimension of observation box}

\item{np}{numpy from package reticulate}

\item{skimage}{skimage}
}
\value{
Vector of vortex properties
}
\description{
Function used to find scar properties given all the pixels in the scar
}
