% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findVorticesCentroids.R
\name{explore.vortex}
\alias{explore.vortex}
\title{Function used to find vortex properties given all the pixels in the vortex}
\usage{
explore.vortex(vortexPixels, boxDimensions)
}
\arguments{
\item{vortexPixels}{Data Frame with coordinates of all the pixels in the}

\item{boxDimensions}{vector of dimension of observation box}
}
\value{
Vector of vortex properties
}
\description{
Function used to find vortex properties given all the pixels in the vortex
}
