% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingFunctions.R
\name{fill_missing}
\alias{fill_missing}
\title{Function for filling in missing values in a image matrix using linear regression
This assumes all the different dx/dy values are multiples of eachother}
\usage{
fill_missing(mat_in, x_in, y_in)
}
\arguments{
\item{mat_in}{Matrix as input}

\item{x_in}{Values connected to the x-indices}

\item{y_in}{Values connected to the y-indices}
}
\description{
Function for filling in missing values in a image matrix using linear regression
This assumes all the different dx/dy values are multiples of eachother
}
