% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surfaceDivergenceManipulation.R
\name{finn.skall.av.vindauga}
\alias{finn.skall.av.vindauga}
\title{Function for finding shell of a given window}
\usage{
finn.skall.av.vindauga(vindauga)
}
\arguments{
\item{vindauga}{Vindauga me ynskjer å finne skallet til}
}
\value{
Ei nx2 matrise med indexar for koordinatar innanfor ein sirkel med den gitte radiusen
}
\description{
Function for finding shell of a given window
}
