% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surfaceDivergenceManipulation.R
\name{finn.vindauga.sirkel}
\alias{finn.vindauga.sirkel}
\title{Function for smoothing out data for some field on the surface}
\usage{
finn.vindauga.sirkel(radius)
}
\arguments{
\item{radius}{Radius man ynskjer å vindauga for}
}
\value{
Ei nx2 matrise med indexar for koordinatar innanfor ein sirkel med den gitte radiusen
}
\description{
Function for smoothing out data for some field on the surface
}
