% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lik_matrices.R
\name{get.fit.specific.r.h}
\alias{get.fit.specific.r.h}
\title{Function for preparing the points, given a list of all points}
\usage{
get.fit.specific.r.h(
  scars_list,
  dimples_list,
  l,
  field,
  tidspunkt,
  r,
  h = 0,
  radiiModulo = 1000L,
  radiiForPrint = -1,
  startLag = 0,
  resType = "aic"
)
}
\arguments{
\item{scars_list}{Complete list of scars for all timesteps}

\item{dimples_list}{Complete list of dimples for all timesteps}

\item{l}{Side length of window}

\item{field}{TxMxN array with surface field}

\item{tidspunkt}{vector with timesteps to use for inference}

\item{r}{radius}

\item{h}{lag}

\item{radiiModulo}{Parameter for the case where we use a single number for both r and h}

\item{radiiForPrint}{Vector of all radii where we should print the radius}

\item{startLag}{Integer, if r is in radiiForPrint, and h == startLag, we print r}

\item{resType}{What type of output is desired? Set to "aic" to get the aic of the fitted model. Set to "fit" for the model object}
}
\description{
Function for preparing the points, given a list of all points
}
