% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point_tracking.R
\name{initialize.point.tracking}
\alias{initialize.point.tracking}
\title{Function for creating beta squared covariates for the desired indices for DNS data}
\usage{
initialize.point.tracking(points.t0)
}
\arguments{
\item{points.t0}{dataframe with points to initialize}
}
\description{
Function for creating beta squared covariates for the desired indices for DNS data
}
