% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surfaceDivergenceManipulation.R
\name{locvar.transformation.periodic}
\alias{locvar.transformation.periodic}
\title{Function for calculating the local variance with periodic boundary conditions}
\usage{
locvar.transformation.periodic(sField, vindauga, skall)
}
\arguments{
\item{sField}{Matrix of values on the surface}

\item{vindauga}{indices of sliding window relative to the pixle we smooth around}

\item{skall}{indices of shell of sliding window relative to the pixle we smooth around}
}
\value{
A smoothed version of sField
}
\description{
Function for calculating the local variance with periodic boundary conditions
}
