% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lik_matrices.R
\name{make.aic.matrix}
\alias{make.aic.matrix}
\title{Function to make an aic matrix using spatstats mppm function}
\usage{
make.aic.matrix(
  scars_list,
  dimples_list,
  l_x,
  l_y,
  field,
  tidspunkt,
  radiar,
  lags,
  radiiModulo = 1000L,
  radiiForPrint
)
}
\arguments{
\item{scars_list}{Complete list of scars for all timesteps}

\item{dimples_list}{Complete list of dimples for all timesteps}

\item{l_x}{Side length of observation window in x direction}

\item{l_y}{Side length of observation window in y direction}

\item{field}{TxMxN array with surface field}

\item{tidspunkt}{vector with timesteps to use for inference}

\item{radiar}{radiar}

\item{lags}{lags}

\item{radiiModulo}{Parameter for the case where we use a single number for both r and h}

\item{radiiForPrint}{Vector of all radii where we should print the radius}
}
\description{
Function to make an aic matrix using spatstats mppm function
}
