% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aic_matrices.R
\name{make.input.matrix}
\alias{make.input.matrix}
\title{Function for making a matrix containing information both on the lag and radius to be used for each index in the aicMatrices}
\usage{
make.input.matrix(radiar, lags, radiiModulo = 1000L)
}
\arguments{
\item{radiar}{Vector with all radii for local variance}

\item{lags}{Vector with all lag values to use}

\item{radiiModulo}{parameter for how to encode lag and radius}
}
\description{
Function for making a matrix containing information both on the lag and radius to be used for each index in the aicMatrices
}
