% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aic_matrices.R
\name{make.new.aic.matrix.experimental}
\alias{make.new.aic.matrix.experimental}
\title{Function for making a new AIC matrix given experimental data}
\usage{
make.new.aic.matrix.experimental(
  radiar,
  hmax,
  vortices.pp,
  scars.pp,
  hDiv,
  indices,
  ensembles,
  printProgress = FALSE,
  skim = 0,
  transformation = "identity"
)
}
\arguments{
\item{radiar}{Vector with all radii for local variance}

\item{hmax}{Maximal lag}

\item{vortices.pp}{votrices as spatstat point process}

\item{scars.pp}{scars as spatstat point process}

\item{hDiv}{horisontal divergence}

\item{indices}{which indices to use}

\item{ensembles}{vector of ensembles to use}

\item{printProgress}{Boolean variable to indicate whether to print progress or not}

\item{skim}{number of pixles to skim of the edge}

\item{transformation}{Should the covariate be transformed, if value is "log" then a log transformation is applied}
}
\description{
Function for making a new AIC matrix given experimental data
}
