% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aic_matrices.R
\name{make.new.aic.matrix}
\alias{make.new.aic.matrix}
\title{Function for making a new AIC matrix given data}
\usage{
make.new.aic.matrix(
  radiar,
  lags,
  vortices.pp,
  scars.pp,
  hDiv,
  indices,
  ensembles = NULL,
  sim = TRUE,
  skim = 0,
  radiiForPrint,
  radiiModulo = 1000L,
  transformation = "identity"
)
}
\arguments{
\item{radiar}{Vector with all radii for local variance}

\item{lags}{Vector with all lag values to use}

\item{vortices.pp}{votrices as spatstat point process}

\item{scars.pp}{scars as spatstat point process}

\item{hDiv}{horisontal divergence}

\item{indices}{which indices to use}

\item{ensembles}{Vector of ensembles to make use of, set to NULL if the data is not split into ensembles}

\item{sim}{Boolean set to true if the data is simulated with periodic BCs}

\item{skim}{Not yet available}

\item{radiiForPrint}{Boolean variable to indicate whether to print progress or not}

\item{radiiModulo}{parameter for encoding lag and radius into single matrix}

\item{transformation}{Should the covariate be transformed, if value is "log" then a log transformation is applied}
}
\description{
Function for making a new AIC matrix given data
}
