% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findVorticesCentroids.R
\name{mean.pixels.crossing.border}
\alias{mean.pixels.crossing.border}
\title{Functions used for going from the complete mapped vortex data, to coordinate data for vortex centroids}
\usage{
\method{mean}{pixels.crossing.border}(pixelCoords, modNum)
}
\arguments{
\item{pixelCoords}{coordinates of pixels in current dimension}

\item{modNum}{Total number of pixels in dimension}
}
\description{
Functions used for going from the complete mapped vortex data, to coordinate data for vortex centroids
}
