% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point_tracking.R
\name{one.step}
\alias{one.step}
\title{Function for tracking points acros a single timestep after some of the initial precautions have been taken care of}
\usage{
one.step(
  points.t1,
  points.t2,
  pointsComplete,
  living,
  t,
  closenessThreshold = 20,
  weight_dist = 256,
  moveOnTorus = TRUE
)
}
\arguments{
\item{points.t1}{dataframe with points at the first of the two timesteps}

\item{points.t2}{dataframe with points at the second of the two timesteps}

\item{pointsComplete}{dataframe to put results into}

\item{living}{matrix with indices of the "living" vortices, first the indices in points.t1, then the related index in pointsComplete}

\item{t}{current timestep}

\item{closenessThreshold}{Parameter to control how far away vortices are allowed to be before they are judged to not be the same}

\item{weight_dist}{parameter to indicate how to weight difference in distance between to points vs the difference in size of two points}

\item{moveOnTorus}{set to TRUE if points move on a torus}
}
\description{
Function for tracking points acros a single timestep after some of the initial precautions have been taken care of
}
