% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grfPeriodicFunctions.R
\name{plotting.transformed.theta}
\alias{plotting.transformed.theta}
\title{Function for plotting the theta values as wanted}
\usage{
plotting.transformed.theta(res, trueParams, kappaUpperLim = NULL)
}
\arguments{
\item{res}{output from parameter estimation}

\item{trueParams}{True parameter values}

\item{kappaUpperLim}{upper limit on the kappa parameter for plotting range}
}
\description{
Function for plotting the theta values as wanted
}
