% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aic_matrices_shrunk_observation.R
\name{point.in.shrunk}
\alias{point.in.shrunk}
\title{Function for checking if any of the points are in the shrunk observation window}
\usage{
point.in.shrunk(points, xlim, ylim)
}
\arguments{
\item{points}{dataframe of points}

\item{xlim}{how much to shrink the window in the x direction}

\item{ylim}{how much to shrink the window in the y direction}
}
\description{
Function for checking if any of the points are in the shrunk observation window
}
