% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aic_matrices.R
\name{prepare.points.dns}
\alias{prepare.points.dns}
\title{Function for preparing the points, given a list of all points}
\usage{
prepare.points.dns(indices, points)
}
\arguments{
\item{indices}{vector of indices/times to use points from}

\item{points}{The complete list of points at all times}
}
\description{
Function for preparing the points, given a list of all points
}
