% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aic_matrices_shrunk_observation.R
\name{prepare.points.dns.shrunk}
\alias{prepare.points.dns.shrunk}
\title{Function for preparing the points, given a list of all points}
\usage{
prepare.points.dns.shrunk(indices, points, xlim, ylim)
}
\arguments{
\item{indices}{vector of indices/times to use points from}

\item{points}{The complete list of points at all times}

\item{xlim}{how much to shrink the window in the x direction}

\item{ylim}{how much to shrink the window in the y direction}
}
\description{
Function for preparing the points, given a list of all points
}
