% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aic_matrices.R
\name{prepare.points.experimental}
\alias{prepare.points.experimental}
\title{Function for preparing the points, given a list of all points}
\usage{
prepare.points.experimental(
  indices,
  numberOfEnsembles,
  points,
  xdim,
  ydim,
  skim = 0
)
}
\arguments{
\item{indices}{indices to use points from}

\item{numberOfEnsembles}{vector with all ensembles to use}

\item{points}{The complete list of points at all times}

\item{xdim}{xdim of covariates used in scaling}

\item{ydim}{ydim of covariates used in scaling}

\item{skim}{how many pixles to skim off the edge?}
}
\description{
Function for preparing the points, given a list of all points
}
