% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{re1000weInf}
\alias{re1000weInf}
\title{Heavily reduced data for DNS data with Reynolds number = 1000, and Weber number = Infinity}
\format{
\subsection{re1000weInf}{

A list with
}
}
\usage{
re1000weInf
}
\description{
Heavily reduced data for DNS data with Reynolds number = 1000, and Weber number = Infinity
}
\keyword{datasets}
