% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aic_matrices_shrunk_observation.R
\name{scale.points.shrunk}
\alias{scale.points.shrunk}
\title{Function for scaling points from a reduced observation window into a "full" observation window}
\usage{
\method{scale}{points.shrunk}(points, xlim, ylim)
}
\arguments{
\item{points}{dataframe of points}

\item{xlim}{how much to shrink the window in the x direction}

\item{ylim}{how much to shrink the window in the y direction}
}
\description{
Function for scaling points from a reduced observation window into a "full" observation window
}
