% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aic_matrices.R
\name{scale.points.skim}
\alias{scale.points.skim}
\title{Function for scaling points to window when skimming edge pixles}
\usage{
\method{scale}{points.skim}(points, c_x, c_y)
}
\arguments{
\item{points}{which points to scale}

\item{c_x}{how to scale x-axis}

\item{c_y}{how to scale y-axis}
}
\description{
Function for scaling points to window when skimming edge pixles
}
