% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgcpPeriodic.R
\name{simu.lgcp.pp}
\alias{simu.lgcp.pp}
\title{Function to simulate a log-Gaussian Cox process with periodic boundary conditions}
\usage{
simu.lgcp.pp(
  discParamX,
  beta,
  theta,
  limX = 1,
  limY = 1,
  discParamY = NULL,
  torus = NULL,
  femMatrices = NULL,
  plotLambda = FALSE,
  plotVortices = FALSE,
  includeLambdaInOutput = TRUE
)
}
\arguments{
\item{discParamX}{Discretization step size in X-direction}

\item{beta}{lambda = exp(beta1+beta2*grf)}

\item{theta}{Parameter values, (tau, kappa)}

\item{limX}{How far to go in x-direction}

\item{limY}{How far to go in y-direction}

\item{discParamY}{Discretization step size in Y-direction. If none is provided, discParamX is used for both X and Y axis}

\item{torus}{Mesh to use for simulation}

\item{femMatrices}{precision matrices, output from fmesher::fm_fem(torus)}

\item{plotLambda}{Set to TRUE if you want a plot of the simulated intensity}

\item{plotVortices}{Set to TRUE if you want a plot of the simulated vortices}

\item{includeLambdaInOutput}{Set to false if you only want the vortices as output}
}
\description{
Function to simulate a log-Gaussian Cox process with periodic boundary conditions
}
