% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grfPeriodicFunctions.R
\name{simu.periodic.grf.2d}
\alias{simu.periodic.grf.2d}
\title{Function to simulate a two dimensional gaussian random field with periodic boundary conditions using the SPDE approach}
\usage{
simu.periodic.grf.2d(
  discParamX,
  theta,
  limX = 1,
  limY = 1,
  discParamY = NULL,
  torus = NULL,
  femMatrices = NULL,
  plotResult = FALSE,
  outputAsDataframe = FALSE
)
}
\arguments{
\item{discParamX}{Discretization step size in X-direction}

\item{theta}{Parameter values, (tau, kappa)}

\item{limX}{How far to go in x-direction}

\item{limY}{How far to go in y-direction}

\item{discParamY}{Discretization step size in Y-direction. If none is provided, discParamX is used for both X and Y axis}

\item{torus}{Mesh to use for simulation}

\item{femMatrices}{precision matrices, output from fmesher::fm_fem(torus)}

\item{plotResult}{Set to TRUE if you want a plot of the resulting}

\item{outputAsDataframe}{If set to true, the output is a dataframe with coordinates and values as columns}
}
\description{
Function to simulate a two dimensional gaussian random field with periodic boundary conditions using the SPDE approach
}
