% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgcpPeriodic.R
\name{simu.pp.from.intensity}
\alias{simu.pp.from.intensity}
\title{Function to simulate inhomogeneous poisson point process}
\usage{
simu.pp.from.intensity(
  lambda,
  discParamX,
  discParamY = NULL,
  limX = 1,
  limY = 1
)
}
\arguments{
\item{lambda}{Intensity of the field}

\item{discParamX}{Discretization step size in X-direction}

\item{discParamY}{Discretization step size in Y-direction. If none is provided, discParamX is used for both X and Y axis}

\item{limX}{How far to go in x-direction}

\item{limY}{How far to go in y-direction}
}
\description{
Function to simulate inhomogeneous poisson point process
}
\examples{
a <- matrix(data = simu.periodic.grf.2d(discParamX = 256,
                                        theta = c(0.1, 5),
                                        plotResult = TRUE), ncol = 256)
lambda <- exp(3*a+2)
summary(as.vector(lambda))
b <- simu.pp.from.intensity(lambda = lambda, discParamX = 256)
plot(b)
image(lambda)

}
