% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findVorticesCentroids.R
\name{skeletonize_object_py}
\alias{skeletonize_object_py}
\title{Skeletonize a scar given pixels}
\usage{
skeletonize_object_py(df, np, skimage)
}
\arguments{
\item{df}{dataframe with the pixel coordinates of the scar}

\item{np}{numpy from package reticulate}

\item{skimage}{skimage}
}
\value{
Centerline of the scar
}
\description{
Skeletonize a scar given pixels
}
