% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surfaceDivergenceManipulation.R
\name{smooth.surface.field}
\alias{smooth.surface.field}
\title{Function for smoothing out data for some field on the surface}
\usage{
smooth.surface.field(sField, vindauga, glatter)
}
\arguments{
\item{sField}{Matrix of values on the surface}

\item{vindauga}{size of slinding window for smoothing}

\item{glatter}{matrix of weights for how to smooth the pixels}
}
\value{
A smoothed version of sField
}
\description{
Function for smoothing out data for some field on the surface
}
