% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lik_matrices.R
\name{surf.transform.specific.t}
\alias{surf.transform.specific.t}
\title{Function for preparing the points, given a list of all points}
\usage{
surf.transform.specific.t(t, field, l, vindauga = NULL, skall = NULL)
}
\arguments{
\item{t}{Timestep to use of array}

\item{field}{TxMxN array with field to be transformed into covariate}

\item{l}{Side length of observation window}

\item{vindauga}{window to use for the calculation of local variance}

\item{skall}{shell of window to be used for calculation of local variance}
}
\description{
Function for preparing the points, given a list of all points
}
