% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingFunctions.R
\name{surface.field.plot}
\alias{surface.field.plot}
\title{Plotting functions
Function for plotting a snapshot of some function/field on the surface}
\usage{
surface.field.plot(sField)
}
\arguments{
\item{sField}{Matrix with surface field data for the interesting snapshot}
}
\description{
Plotting functions
Function for plotting a snapshot of some function/field on the surface
}
