% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingFunctions.R
\name{surface.field.pp.plot}
\alias{surface.field.pp.plot}
\title{Function for plotting a snapshot of surface field and surface pp together}
\usage{
surface.field.pp.plot(sField, sPPCoords)
}
\arguments{
\item{sField}{Matrix with surface field data for the interesting snapshot}

\item{sPPCoords}{Coordinates for the point process}
}
\description{
Function for plotting a snapshot of surface field and surface pp together
}
