% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{surfaceDivergence}
\alias{surfaceDivergence}
\title{Numerically Simulated Surface Divergence data
Generated using the command:
library("rhdf5")
surfaceDivergence <- h5read(file = '/Users/danielks/Library/CloudStorage/OneDrive-NTNU/PhD/SimData/data/surface_vortices.h5', name = 'sdiv')}
\format{
\subsection{surfaceDivergence}{

A large 3-dimensional array with 12500 timesteps of surface divergence in a 256-by-256 grid for every timestep
}
}
\usage{
surfaceDivergence
}
\description{
Numerically Simulated Surface Divergence data
Generated using the command:
library("rhdf5")
surfaceDivergence <- h5read(file = '/Users/danielks/Library/CloudStorage/OneDrive-NTNU/PhD/SimData/data/surface_vortices.h5', name = 'sdiv')
}
\keyword{datasets}
