% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingFunctions.R
\name{surfacePP.plot}
\alias{surfacePP.plot}
\title{Function for plotting a snapshot of some point process on the surface}
\usage{
surfacePP.plot(sPPCoords)
}
\arguments{
\item{sPPCoords}{Coordinates for the point process}
}
\description{
Function for plotting a snapshot of some point process on the surface
}
