% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{surfaceVortices}
\alias{surfaceVortices}
\title{Numerically Simulated Surface Vortex data
Generated using the commands:
library("rhdf5")
surfaceVortices <- h5read(file = '/Users/danielks/Library/CloudStorage/OneDrive-NTNU/PhD/SimData/data/surface_vortices.h5', name = 'surf')}
\format{
\subsection{surfaceDivergence}{

A large 3-dimensional array with 12500 timesteps in a 256-by-256 grid for every timestep every pixel is either a one or a zero, with a one if there is a vortex, and a zero if there is not.
}
}
\usage{
surfaceVortices
}
\description{
Numerically Simulated Surface Vortex data
Generated using the commands:
library("rhdf5")
surfaceVortices <- h5read(file = '/Users/danielks/Library/CloudStorage/OneDrive-NTNU/PhD/SimData/data/surface_vortices.h5', name = 'surf')
}
\keyword{datasets}
