% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point_tracking.R
\name{track.points.over.time}
\alias{track.points.over.time}
\title{Function for tracking points moving in some observation window}
\usage{
track.points.over.time(
  pointsUntracked,
  tEnd = NULL,
  tStart = 1,
  closenessThreshold = 20,
  weight_dist = 256,
  moveOnTorus = TRUE
)
}
\arguments{
\item{pointsUntracked}{A list of snapshots of points, each point on a snapshot must contain at least, x, y and noPixels}

\item{tEnd}{time to end}

\item{tStart}{time to start, if no input is given, this will be set to the length of pointsUntracked}

\item{closenessThreshold}{Parameter to control how far away vortices are allowed to be before they are judged to not be the same}

\item{weight_dist}{parameter to indicate how to weight difference in distance between to points vs the difference in size of two points}

\item{moveOnTorus}{set to TRUE if points move on a torus}
}
\description{
Function for tracking points moving in some observation window
}
