% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingFunctions.R
\name{aicPlot}
\alias{aicPlot}
\title{Function for plotting a snapshot of surface field and surface pp together}
\usage{
aicPlot(
  aicMatrix,
  lagValues,
  timeScale,
  lengthScale,
  radiar,
  plotAsContour = FALSE
)
}
\arguments{
\item{aicMatrix}{Matrix with surface field data for the interesting snapshot}

\item{lagValues}{Coordinates for the point process}

\item{timeScale}{how many timescales per frame?}

\item{lengthScale}{how many lengthscales per pixle?}

\item{radiar}{description}

\item{plotAsContour}{set to TRUE if you want a contour plot instead of a heatmap}
}
\description{
Function for plotting a snapshot of surface field and surface pp together
}
