% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findVorticesCentroids.R
\name{find.scars.centrelines.specific.time}
\alias{find.scars.centrelines.specific.time}
\title{Function to go from complete mapped scar pixel data to the point process format of scar data}
\usage{
find.scars.centrelines.specific.time(scarData, np = NULL, skimage = NULL)
}
\arguments{
\item{scarData}{Matrix of zeros and ones, where the ones are the pixels with a scar on them}

\item{np}{numpy from package reticulate}

\item{skimage}{skimage}
}
\value{
Dataframe with scar data
}
\description{
Function to go from complete mapped scar pixel data to the point process format of scar data
}
