% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surfaceDivergenceManipulation.R
\name{smooth.surface.field}
\alias{smooth.surface.field}
\title{Function for smoothing out data for some field on the surface}
\usage{
smooth.surface.field(sField, vindauga, glatter, boundaryCondition)
}
\arguments{
\item{sField}{Matrix of values on the surface}

\item{vindauga}{indices of sliding window relative to the pixle we smooth around}

\item{glatter}{matrix of weights for how to smooth the pixels}

\item{boundaryCondition}{"periodic" if we have periodic data, "physical" if not}
}
\value{
A smoothed version of sField
}
\description{
Function for smoothing out data for some field on the surface
}
